requirements.txt
openai
elevenlabs
moviepy
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
ffmpeg-python
